ELSEPTIMOREPOSITORIO-YAAAAA--main/ELSEPTIMOREPOSITORIO-YAAAAA--main/Nasa-Back-main/requirements.txt
Flask
python-dotenv
google-generativeai